Completely remove old non-structured errors and migrate existing ones.
