--- app/policies/code_ocean/error_policy.rb
@@ -1,7 +1,0 @@
-module CodeOcean
-  class ErrorPolicy < AdminOrAuthorPolicy
-    def author?
-      @user == @record.execution_environment.author
-    end
-  end
-end
