--- app/views/code_ocean/errors/show.html.slim
@@ -1,4 +1,0 @@
-h1 = CodeOcean::Error.model_name.human
-
-= row(label: 'error.message', value: @error.message)
-= row(label: 'shared.created_at', value: l(@error.created_at, format: :short))
