+++ app/views/execution_environments/index.html.slim
@@ -11,7 +11,6 @@ h1 = ExecutionEnvironment.model_name.human(count: 2)
         th = t('activerecord.attributes.execution_environment.network_enabled')
         th = t('activerecord.attributes.execution_environment.permitted_execution_time')
         th colspan=5 = t('shared.actions')
-        th colspan=2 = t('shared.resources')
     tbody
       - @execution_environments.each do |execution_environment|
         tr
@@ -26,7 +25,6 @@ h1 = ExecutionEnvironment.model_name.human(count: 2)
           td = link_to(t('shared.destroy'), execution_environment, data: {confirm: t('shared.confirm_destroy')}, method: :delete)
           td = link_to(t('.shell'), shell_execution_environment_path(execution_environment))
           td = link_to(t('shared.statistics'), statistics_execution_environment_path(execution_environment))
-          td = link_to(t('activerecord.models.error.other'), execution_environment_errors_path(execution_environment.id))
 
 = render('shared/pagination', collection: @execution_environments)
 p = render('shared/new_button', model: ExecutionEnvironment)
