+++ db/seeds/development.rb
@@ -10,9 +10,6 @@ Rails.application.routes.default_url_options = Rails.application.config.action_m
 # execution environments
 ExecutionEnvironment.create_factories
 
-# errors
-CodeOcean::Error.create_factories
-
 # exercises
 @exercises = find_factories_by_class(Exercise).map(&:name).map { |factory_name| [factory_name, FactoryBot.create(factory_name)] }.to_h
