--- spec/factories/code_ocean/error.rb
@@ -1,6 +1,0 @@
-FactoryBot.define do
-  factory :error, class: CodeOcean::Error do
-    association :execution_environment, factory: :ruby
-    message { "exercise.rb:4:in `<main>': undefined local variable or method `foo' for main:Object (NameError)" }
-  end
-end
